fixed the code to save to loal storage and show only arrays
