+++ src/tempCodeRunnerFile.js
@@ -1,0 +1,4 @@
+const tasks = [];
+const taskIndex = tasks.length;
+const listItem = document.createElement('li');
+listItem.setAttribute('data-index', taskIndex);
